# Conflicts:
#	xmipp3/__init__.py

+++ xmipp3/viewers/viewer_resolution_monogenic_signal.py
@@ -176,7 +176,7 @@ class XmippMonoResViewer(LocalResolutionViewer):
         sliceNumber = self.sliceNumber.get()
         if sliceNumber < 0:
             x ,_ ,_ ,_ = ImageHandler().getDimensions(imageFile)
-            sliceNumber = x/2
+            sliceNumber = int(x/2)
         else:
             sliceNumber -= 1
         #sliceNumber has no sense to start in zero
